Moved sparql generation functionality from RDFModelUtil to new class
