+++ fcrepo-clients/src/main/java/edu/unc/lib/dl/fcrepo4/RepositoryObjectFactory.java
@@ -42,6 +42,7 @@ import edu.unc.lib.dl.fedora.PID;
 import edu.unc.lib.dl.rdf.Cdr;
 import edu.unc.lib.dl.rdf.PcdmModels;
 import edu.unc.lib.dl.rdf.Premis;
+import edu.unc.lib.dl.sparql.SparqlUpdateHelper;
 import edu.unc.lib.dl.sparql.SparqlUpdateService;
 import edu.unc.lib.dl.util.RDFModelUtil;
 import edu.unc.lib.dl.util.URIUtil;
@@ -347,7 +348,7 @@ public class RepositoryObjectFactory {
             // If a model was provided, then add the triples to the new binary's
             // metadata
             // Turn model into sparql update query
-            String sparqlUpdate = RDFModelUtil.createSparqlInsert(model);
+            String sparqlUpdate = SparqlUpdateHelper.createSparqlInsert(model);
             InputStream sparqlStream = new ByteArrayInputStream(sparqlUpdate.getBytes(StandardCharsets.UTF_8));
 
             try (FcrepoResponse response = getClient().patch(describedBy).body(sparqlStream).perform()) {
@@ -399,7 +400,7 @@ public class RepositoryObjectFactory {
      * @param object
      */
     public void createProperty(PID subject, Property property, String object) {
-        String sparqlUpdate = RDFModelUtil.createSparqlInsert(subject.getRepositoryPath(), property, object);
+        String sparqlUpdate = SparqlUpdateHelper.createSparqlInsert(subject.getRepositoryPath(), property, object);
         persistTripleToFedora(subject, sparqlUpdate);
     }
 
@@ -425,7 +426,7 @@ public class RepositoryObjectFactory {
      * @param object
      */
     public void createRelationship(PID subject, Property property, Resource object) {
-        String sparqlUpdate = RDFModelUtil.createSparqlInsert(subject.getRepositoryPath(), property, object);
+        String sparqlUpdate = SparqlUpdateHelper.createSparqlInsert(subject.getRepositoryPath(), property, object);
         persistTripleToFedora(subject, sparqlUpdate);
     }
 
@@ -435,7 +436,7 @@ public class RepositoryObjectFactory {
      * @param model
      */
     public void createRelationships(PID pid, Model model) {
-        String sparqlUpdate = RDFModelUtil.createSparqlInsert(model);
+        String sparqlUpdate = SparqlUpdateHelper.createSparqlInsert(model);
         persistTripleToFedora(pid, sparqlUpdate);
     }
