+++ metadata/src/main/java/edu/unc/lib/dl/util/RDFModelUtil.java
@@ -22,16 +22,8 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
-import org.apache.jena.datatypes.RDFDatatype;
-import org.apache.jena.datatypes.TypeMapper;
-import org.apache.jena.graph.Node;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ModelFactory;
-import org.apache.jena.rdf.model.Property;
-import org.apache.jena.rdf.model.ResIterator;
-import org.apache.jena.rdf.model.Resource;
-import org.apache.jena.rdf.model.Statement;
-import org.apache.jena.rdf.model.StmtIterator;
 import org.apache.jena.riot.RDFDataMgr;
 import org.apache.jena.riot.RDFFormat;
 
@@ -47,7 +39,6 @@ public class RDFModelUtil {
     public final static String TURTLE_MIMETYPE = "text/turtle";
 
     private RDFModelUtil() {
-
     }
 
     public static void serializeModel(Model model, File file) throws IOException {
@@ -94,174 +85,5 @@ public class RDFModelUtil {
         return model;
     }
 
-    /**
-     * Convert the given model into a SPARQL update query which inserts all of
-     * the properties in the model.
-     *
-     * @param model
-     * @return sparql update query which adds all properties from the given
-     *         model
-     */
-    public static String createSparqlInsert(Model model) {
-        StringBuilder query = new StringBuilder();
-        query.append("INSERT {\n");
 
-        ResIterator it = model.listSubjects();
-        while (it.hasNext()) {
-            Resource resc = it.nextResource();
-            String currentUri = resc.getURI();
-
-            StmtIterator pIt = resc.listProperties();
-            while (pIt.hasNext()) {
-                Statement property = pIt.nextStatement();
-                query.append(" <").append(currentUri).append('>').append(" <")
-                        .append(property.getPredicate().getURI()).append("> ");
-                if (property.getObject().isResource()) {
-                    query.append('<')
-                            .append(property.getObject().asResource().getURI())
-                            .append('>');
-                } else {
-                    Node node = property.getObject().asNode();
-                    query.append('"').append(node.getLiteralLexicalForm())
-                            .append('"');
-                    String typeUri = node.getLiteralDatatypeURI();
-                    if (typeUri != null) {
-                        query.append("^^<")
-                                .append(node.getLiteralDatatypeURI())
-                                .append('>');
-                    }
-                }
-                query.append(" .\n");
-            }
-        }
-
-        query.append("}\nWHERE {}");
-
-        return query.toString();
-    }
-
-    /**
-     * Create a SPARQL update query to add one property to a resource
-     *
-     * @param subjUri
-     * @param property
-     * @param object
-     * @return
-     */
-    public static String createSparqlInsert(String subjUri, Property property,
-            Object object) {
-
-        String objectString = getObjectAsString(object);
-
-        StringBuilder query = new StringBuilder();
-        addInsert(query, subjUri, property, objectString);
-        query.append("\nWHERE {}");
-
-        return query.toString();
-    }
-
-    /**
-     * Constructs a sparql query which replaces all instances of a property on a resource with a new valu
-     *
-     * @param subjUri
-     * @param property
-     * @param object
-     * @return
-     */
-    public static String createSparqlReplace(String subjUri, Property property,
-            Object object) {
-
-        String objectString = getObjectAsString(object);
-
-        String subjString = getSubjectString(subjUri);
-
-        StringBuilder query = new StringBuilder();
-        query.append("DELETE { ");
-        query.append(subjString).append(" <").append(property.getURI()).append("> ?obj . }\n");
-
-        addInsert(query, subjUri, property, objectString);
-
-        query.append("\nWHERE {");
-        query.append(subjString).append(" <").append(property.getURI()).append("> ?obj . }");
-
-        return query.toString();
-    }
-
-    /**
-     * Constructs a sparql update query to delete all instances of a property from a resource.
-     *
-     * @param subjUri
-     * @param property
-     * @return
-     */
-    public static String createSparqlDelete(String subjUri, Property property, Object object) {
-        String subjString = getSubjectString(subjUri);
-
-        StringBuilder query = new StringBuilder();
-        query.append("DELETE { ");
-        query.append(subjString).append(" <").append(property.getURI()).append('>');
-
-        // If no object was specified for deletion, then select for all values of property
-        if (object == null) {
-            query.append(" ?obj . }\n");
-        } else {
-            query.append(' ').append(getObjectAsString(object)).append(" . }\n");
-        }
-
-        query.append("\nWHERE {");
-        if (object == null) {
-            query.append(subjString).append(" <").append(property.getURI()).append("> ?obj . }");
-        } else {
-            query.append('}');
-        }
-
-        return query.toString();
-    }
-
-    /**
-     * Returns a string representation of a resource or a typed literal based on
-     * the type of object provided. If the object is not a known XSD datatype or
-     * resource, null is returned.
-     *
-     * @param object
-     * @return
-     */
-    public static String getObjectAsString(Object object) {
-        if (object instanceof Resource) {
-            return '<' + ((Resource) object).getURI() + '>';
-        } else if (object instanceof String) {
-            return '"' + object.toString() + '"';
-        } else {
-            RDFDatatype type = TypeMapper.getInstance().getTypeByClass(object.getClass());
-            if (type == null) {
-                throw new IllegalArgumentException("Cannot convert object of type " + object.getClass()
-                        + ", only supported XSD datatypes and Resources are supported.");
-            }
-            String typeUri = type.getURI();
-
-            return '"' + object.toString() + "\"^^<" + typeUri + ">";
-        }
-    }
-
-    private static void addInsert(StringBuilder query, String subjUri, Property property,
-            String object) {
-        query.append("INSERT {");
-
-        if (subjUri == null) {
-            query.append(" <>");
-        } else {
-            query.append(" <").append(subjUri).append('>');
-        }
-
-        query.append(" <").append(property.getURI()).append("> ")
-                .append(object).append(" . ").append("}");
-    }
-
-    private static String getSubjectString(String subjUri) {
-        if (subjUri == null) {
-            return "<>";
-        } else {
-            return "<" + subjUri + ">";
-        }
-    }
 }
